Classes & Functions
Two classes were created. One is called "Items" and represents the products. The other is called "Storage" and represents any generic kind of storage (e.g. cart and store are Storage instances).
The UX was improved by adding functions who's execute when onclick events are listened. This functions allows to interact with stock property of the items inside the storages.